+++ src/main/java/insane96mcp/progressivebosses/module/dragon/phase/CrystalRespawnPhase.java
@@ -143,6 +143,8 @@ public class CrystalRespawnPhase extends AbstractDragonPhaseInstance {
 		spikes.sort(Comparator.comparingInt(SpikeFeature.EndSpike::getRadius));
 		int spawned = 0;
 		for (SpikeFeature.EndSpike spike : spikes) {
+			if (dragon.level().getEntitiesOfClass(EndCrystal.class, spike.getTopBoundingBox()).isEmpty())
+				continue;
 			this.addCrystalRespawn(spike);
 			if (++spawned >= crystalsToRespawn)
 				break;